Moved RootLayout to its own component, PluginProvider now handle RootProvider and main LayoutDiv
