+++ src/index.js
@@ -1,2 +1,4 @@
 export LayoutState from './model/LayoutState';
+export Provider from './components/Provider';
+export RootLayout from './components/RootLayout';
 export Layout from './Layout';
